+++ CHANGES.rst
@@ -29,6 +29,18 @@
   See `issue 156
   <https://github.com/zopefoundation/BTrees/issues/156>`_.
 
+- Make the BTree objects (``BTree``, ``TreeSet``, ``Set``, ``Bucket``)
+  of each module actually provide the interfaces defined in
+  ``BTrees.Interfaces``. Previously, they provided no interfaces.
+
+- Update the definitions of ``ISized`` and ``IReadSequence`` to simply
+  be ``zope.interface.common.collections.ISized`` and
+  ``zope.interface.common.sequence.IMinimalSequence`` respectively.
+
+- Remove the ``__nonzero__`` interface method from ``ICollection``. No
+  objects actually implemented such a method; instead, the boolean value
+  is typically taken from ``__len__``.
+
 4.7.2 (2020-04-07)
 ==================